# lib/api-zod/src/generated/types/requestUploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestUploadUrlResponseMetadata } from "./requestUploadUrlResponseMetadata";

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlResponseMetadata;
}